Implementation of running ShellTask with workgraph (#45)
Initialization of all files required to run workflows. The workflow config files
are updated to work properly. The $PWD environment is removed because the
`config_rootdir` is now used.

To allow a graph like initialization of workgraph tasks first creating the task
nodes and then linking the additional information including inputs, outputs and
arguments we implemented a temporary hack to overcome internal validations.