Add option to specify computer for available data items (#79)
This allows to specify data items that are only remotely available or should
not be copied over to the aiida internal folder. I did not add the computer
information to the printing of the IR because the IR focuses on representing
the unrolling/expansion from the config to a graph and the computer information
is completely independent from this logic.