Moves logic from `CanonicalWorkflow` to `core.Workflow` and `ConfigWorkflow`
The validation of `cycles` and `tasks` in `list_not_empty` is moved to the
`ConfigWorkflow`. This subsequently required a modification of the tests that
initalize a `ConfigWorkflow` with empty `cycles` and `tasks` as this is not
anymore allowed through the tests. We switched to `BeforeValidator`
as the check can be happen before pydantic validation.

The `data_dict` and `task_dict` member variables are moved to the
constructor of the `core.Workflow`. The constructor of `core.Workflow`
is split into two: `from_config_workflow` constructor, which replicates
the behavir of the previous constructor and the default construcor now
accepts all required that are extracted from the passed `ConfigWorkflow`
in the previous constructor as individual parameters.

The introduction of the `from_config_file` constructor in ConfigWorkflow, which
replicates the behavior of `load_workflow_config`, allows the `rootdir` to
become a part of `ConfigWorkflow`. This approach eliminates the need for an
external utility function for workflow creation from a file. Since the utility
function was essentially acting as a constructor for `ConfigWorkflow`, it
simplifies the interface by enabling direct access to the `rootdir` within
`core.Workflow`. In addition we make `rootdir` and `name` not optional as these
parameters cane determined in the new `from_config_file` and passed to the
default constructor. Furthermore, We introduce the util function
`validate_yaml_content` that allows a more generic usage of creating an instance
from a `Conig*` class from a yaml string, especially used in the docstring
tests.