Remove `pydantic-yaml` as dependency (#110)
The amount of functionality we use of this library is extremly minimal.
We replace it with the dependency `ruamel.yaml` that parses the yaml
file into a python object. Furthermore, the library still uses pydantic
API v1 for validating the pydantic model which is deprecated so we can
replace it with pydantic v2. In addition this change gives us more
flexibility as it separates the parsing of the yaml file and the
validation through pydantic.

To have a similar utility function that can be used in the docstring
tests we implement `validate_yaml_content`.