Rework command line specification (#125)
This PR introduces the "{PORT::port_name}" placeholder in the shell task command and thus extends the port concept to all tasks. This allows for arbitrary command lines to be specified without having to make assumptions when parsing it.

The port keyword is now mandatory in the yaml config file and inputs are now internally represented by a dictionary mapping port names to the corresponding input data nodes list.

Some minor changes are also made to workgraph.py for readability.