Remove validation of src for AvailableData
I decided to remove the check since we will need to move it to to the
aiida part after PR #136 is merged. With #136 we need to consider remote
data which we cannot check neither in core nor in parsing. So moving this
now to the yaml parsing would require the adoption of many test which
will be anyway not be needed.