Fix remote data copying/symlinking with dates and parameters (#157)
Symlink files between tasks using unique labels

With this PR, the `filenames` argument is being set on shell tasks such
that full input data labels are being used, including time stamps and
parametrization information (if applicable).

For instance, for analysis/post-processing tasks, rather than symlinking
every `icon_output` file of previous ICON runs to an `icon_output`
target in the working directory of the analysis/post-processing task,
which would lead to conflicts or silent overwriting the full label
including the time stamp and parametrization information is being used
as the symlink target, ensuring the information is linked correctly.

For this, consistent labels need to be also used in the `nodes` argument
to the shell task (this was already the case), as well as the
`arguments` argument. The functionality is implemented by modifying the
`_set_shelljob_arguments` to use the full label as placeholder, and
introducing a new `_set_shelljob_filenames` method to set the filenames.

The `test_shell_filenames_nodes_arguments` test contains the explicit
`filenames`, `nodes`, and `arguments` set for a shortened version of the
`parameters` config case.

Co-authored-by: Alexander Goscinski <alex.goscinski@posteo.de>