Improve error handling when secrets are not accessible (#156)
When the secrets for spack-buildcache is not accessible, the error message is
not transparent about the exact reason it failed. To prevent such an
intransparent message we check in the integration workflow if the PR is merged
from a fork. In this case the secret is not accessible and we exit the workflow
earlier with a meaningful error message.