Enforce usage of absolute path and computer (#153)
To be more explicit in the config files on which computer the file lies we
enforce the usage of absolute path everywhere except `GenerateData` and
`NamelistFile`. The `computer` argument is now required for tasks as it
specifies where the executable lies. 

Because the config files need to now specify absolute paths, a logic to support
environmental variables in the config file and expand them is implemented.