Move expansion of environment variables to tests (#171)
To simplify the code we remove the functionality to expand environment variables
in the code.
    
For the tests, some kind of mechanism of a variable in the config that is
replaced is required as the tests config need to change depending on the
development environment. The absolute tests directory changes for different
development environment and can thus not be hardcoded. The config requires
absolute path as this is part of the config design. Before this PR environment
variables in the paths have been expanded. This expansion is replaced in this PR
by the pytest fixture `config_paths` that before passing the file to sirocco
replaces the string `/TESTS_ROOTDIR` in the config file by the actual tests root
directory. This mechanism requires a modification of the config files. To
prevent touching the files in the tests directory, test cases are copied to a
tmp path and the replacement is there performed.

Validations if the paths are absolute are reintroduced into the yaml parsing.
With the introduction of the environment variables, these had to be removed
because one had to consider that an environment variable needs to be resolved
during workflow-runtime. An path of the form `$TEST_ROOTDIR/path` is considered
as relative and thus could not be validated.  The check for `ShellTask.src` is
skipped because it will be enforced to be relative in PR #172.