Support copy of local scripts to remote machine (#95)
Support the copy of local scripts to remote machine by adding the script
specified in `src` to the nodes of the `ShellJob`. Because aiida does
not copy the file over preserving the execution permission, we cannot
support a syntax of the form `./icon.py`. Therefore I changed the usage
to the syntax "python icon.py".