Implement MPI and scheduler options (#190)
The mpirun command can be specified in the field `mpi_cmd`. Also support for
`walltime`, `nodes`, `ntasks_per_node` and `cpus_per_task` is implemented. We
introduce the tag `MPI_TOTAL_PROCS` that allows to automatically use the total
number of mpi procs given from the nodes and `ntasks_per_node`.

We have to bound aiida <2.7.0 since there is a transport upload problem in the
CI.

---------

Co-authored-by: Alexander Goscinski <alex.goscinski@posteo.de>